requirements.txt
# Usage: pip install -r requirements.txt
tensorflow
termcolor
pickel
base64
pillow
streamlit

